--- src/components/Icone/Icone.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Icone(props) {
-  return (
-    <div>
-      <img alt={"Icone"} src={props.icone} onClick={props.onClickIcone} />
-    </div>
-  );
-}
-
-export default Icone;
